+++ src/Macros/ResponseTransformer/CreateFrom.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Apiato\Core\Macros\ResponseTransformer;
+
+use Apiato\Core\Services\ResponseTransformer;
+use League\Fractal\Serializer\SerializerAbstract;
+use League\Fractal\TransformerAbstract;
+
+class CreateFrom {
+    public function __invoke(): callable
+    {
+        return function (mixed $data = null, callable|TransformerAbstract|null $transformer = null, SerializerAbstract|null $serializer = null): ResponseTransformer {
+            return ResponseTransformer::create($data, $transformer, $serializer);
+        };
+    }
+}
