+++ src/Macros/ResponseTransformer/Created.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace Apiato\Core\Macros\Fractal;
+namespace Apiato\Core\Macros\ResponseTransformer;
 
 use Apiato\Core\Abstracts\Transformers\Transformer;
 use Illuminate\Http\JsonResponse;
-use Apiato\Core\Services\Response;
+use Apiato\Core\Services\ResponseTransformer;
 
-class Ok {
+class Created {
     public function __invoke(): callable
     {
         return function (): JsonResponse {
-            /** @var Response $this */
+            /** @var ResponseTransformer $this */
             if (is_null($this->getTransformer())) {
                 $this->transformWith(Transformer::empty());
             }
-            return $this->respond(200);
+            return $this->respond(201);
         };
     }
 }
