+++ src/Macros/ResponseTransformer/GetTransformer.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Apiato\Core\Macros\Fractal;
+namespace Apiato\Core\Macros\ResponseTransformer;
 
-use Apiato\Core\Services\Response;
+use Apiato\Core\Services\ResponseTransformer;
 use League\Fractal\TransformerAbstract;
 
 class GetTransformer
@@ -10,7 +10,7 @@ class GetTransformer
     public function __invoke(): callable
     {
         return function (): string|callable|TransformerAbstract|null {
-            /** @var Response $this */
+            /** @var ResponseTransformer $this */
             // The warning is false positive. We will be in the context of Fractal class when this is called.
             return $this->transformer;
         };
