+++ src/Macros/ResponseTransformer/NoContent.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Apiato\Core\Macros\Fractal;
+namespace Apiato\Core\Macros\ResponseTransformer;
 
 use Apiato\Core\Abstracts\Transformers\Transformer;
 use Illuminate\Http\JsonResponse;
-use Apiato\Core\Services\Response;
+use Apiato\Core\Services\ResponseTransformer;
 
 class NoContent {
     public function __invoke(): callable
     {
         return function (): JsonResponse {
-            /** @var Response $this */
+            /** @var ResponseTransformer $this */
             if (is_null($this->getTransformer())) {
                 $this->transformWith(Transformer::empty());
             }
